Get principal from plug and improve ergonomics (#24)
* fix: AccessToken -> ApplicationToken

Cloudflare Acccess specifically calls the type of token we are verifying
"application tokens" in their docs. This commit renames the
AccessTokenVerifier to ApplicationTokenVerifier to match the terminology
and any other references to "access token" to "application token".

https://developers.cloudflare.com/cloudflare-one/identity/authorization-cookie/application-token/

* feat: plug stores user principal

Allowing users of the plug to access the principal of the user
after it has been verified.

* feat: create a principal struct

The representation of the current user from AccessTokenVerifier
felt a little awkward, so I've created a Principal struct to
represent the current user. This also allows us to add more
information to the struct in the future if we need to.